Updated the arquitecture of the application
