+++ SmartLock/templates/index.html
@@ -9,8 +9,8 @@
 
 <form class="box" action="http://localhost:5000/login" method="post">
   <h1>Pi Login</h1>
-  <input type="text" name="" placeholder="Username">
-  <input type="password" name="" placeholder="Password">
+  <input type="text" name="username" placeholder="Username">
+  <input type="password" name="password" placeholder="Password">
   <input type="submit" name="" value="Login">
 </form>
